gdal_footprint/app/requirements.txt
fastapi>=0.85.0
uvicorn>=0.19.0
python-multipart>=0.0.6
gdal>=3.6.0
psycopg2-binary>=2.9.5
pydantic>=1.10.2
python-dotenv>=0.21.0